requirements.txt
anyio==3.7.0
autoflake==1.7.8
black==23.3.0
certifi==2023.5.7
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
fastapi==0.98.0
filelock==3.12.2
flake8==5.0.4
flake8-broken-line==1.0.0
flake8-isort==6.0.0
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
identify==2.5.24
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.1
pathspec==0.11.1
pep8-naming==0.13.3
platformdirs==3.8.0
pluggy==1.2.0
pre-commit==3.3.3
pycodestyle==2.9.1
pydantic==1.10.10
pyflakes==2.5.0
pytest==7.4.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.0
virtualenv==20.23.1
